Added title translation
Title is translated into English using the translate library in order to fill out the trans-title parameter. While the source language could be determined by inspecting the lang HTML attribute, the FastText library is currently used to predict it. Now the goal is to have the user be able to select the destination language (and probably also the source language to override the FastText prediction) in the web front end.