Extract website using tldextract
Website name (domain and suffix) is extracted from the URL using tldextract if a site name isn't specified in the metadata. In addition to this, some minor changes have been made.
If the source language of the title isn't different from the target language, the language and trans-title params aren't included in the reference.
WORK and LOCALE attributes have been renamed to 'WEBSITE' and 'LANGUAGE', respectively.